Blocked plugin when license invalid, removed create repository blocking as it's not compatible with atlassian policy

+++ src/main/java/org/networkedassets/atlassian/stash/privaterepos/repositories/listeners/RepositoryDeletedListener.java
@@ -3,6 +3,7 @@ package org.networkedassets.atlassian.stash.privaterepos.repositories.listeners;
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
 
+import org.networkedassets.atlassian.stash.privaterepos.license.LicenseManager;
 import org.networkedassets.atlassian.stash.privaterepos.repositories.PersonalRepositoriesService;
 import org.networkedassets.atlassian.stash.privaterepos.repositories.RepositoryTypeVerifier;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,6 +23,8 @@ public class RepositoryDeletedListener {
 	private RepositoryTypeVerifier repositoryTypeVerifier;
 	@Autowired
 	private EventPublisher eventPublisher;
+	@Autowired
+	private LicenseManager licenseManager;
 
 	@PostConstruct
 	public void registerEvents() {
@@ -35,6 +38,9 @@ public class RepositoryDeletedListener {
 
 	@EventListener
 	public void handleDeletionEvent(RepositoryDeletedEvent event) {
+		if (licenseManager.isLicenseInvalid()) {
+			return;
+		}
 		Repository repo = event.getRepository();
 		if (repositoryTypeVerifier.isPersonal(repo)) {
 			personalRepositoriesService.deletePersonalRepository(repo);
